fastapi_template/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
flower==2.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2
transformers==4.36.2
torch==2.1.1
pandas==2.1.4
numpy==1.24.4
cryptography==42.0.8
python-dotenv==1.0.0
loguru==0.7.2
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0